moved window state related declarations to separate file
