+++ internal/tui/models/state/state.go
@@ -87,34 +87,3 @@ func (avs *AppPageState) ShowPage(pageId models.PageId) (tea.Cmd, error) {
 
 	return page.InitOnce(), nil
 }
-
-type AppWindowStater interface {
-	GetWindowHeight() int
-	GetWindowWidth() int
-	SetWindowHeight(int)
-	SetWindowWidth(int)
-}
-
-type AppWindowState struct {
-	width, height int
-}
-
-func NewAppWindowState() *AppWindowState {
-	return &AppWindowState{}
-}
-
-func (aws *AppWindowState) GetWindowHeight() int {
-	return aws.height
-}
-
-func (aws *AppWindowState) GetWindowWidth() int {
-	return aws.width
-}
-
-func (aws *AppWindowState) SetWindowHeight(h int) {
-	aws.height = h
-}
-
-func (aws *AppWindowState) SetWindowWidth(w int) {
-	aws.width = w
-}
